enhanced options handling (#7)
* more complete options handling

The existing code was lacking the ability for restored classifier to retain options.

* more complete options handling

This allows custom options to be used during training. I also created a new "top"option that if given in the classifier options it will be used to determine how many of the most probable labels to return. It returns them as and array of objects (containing label and value) similar to getClassifications().

This is very useful if you need to know how confidently the results should be treated.

* fix test complaints

fix test complaints

* fix test to work with new option enhancement

* fix test complaints

* removing classifyTop()